fix: rename internal update_pint_locale to _update_pint_locale_impl
Resolves naming conflict between public API wrapper and internal implementation.

Changes:
- Renamed localization/pint_locale.py:update_pint_locale -> _update_pint_locale_impl
- Updated wrapper in pint_sympy.py to call renamed internal function
- Fixed all test imports to use public API from pint_sympy
- Tests now use same interface as users (no unitregistry parameter)

Benefits:
- Clear API boundaries: public wrapper vs internal implementation
- Prevents accidental direct calls to internal function
- Tests verify the same interface documented for users
- Follows Python naming conventions (_impl suffix for internal code)

All 27 tests in test_localization.py passing.
